// angular/src/app/model/FIneResult.ts
export class FineResult {
  companyName: string = "";
  pollutionName: string = "";
  pollutionValue: number = -1;
  taxForStationarySources: number = 0;
  taxForDischargesToWaterBodies: number = 0;
  taxForWasteDisposal: number = 0;
  taxForTheGenerationOfRadioactiveWaste: number = 0;
  taxForTemporaryStorageOfRadioactiveWaste: number = 0;
  year: string = "";
}